SmartNav: Reload icons from storage when boot completes
This is necessary for the outlying usecase where icons from
storage are set and the device is encrypted. The navbar may have
been created before the device became decrypted thus not being able
to load icons from storage. In the future, it would be better to
respond to a acutal callback when device becomes decrypted

Change-Id: Ic8e711aebc4089ecf1e69e226fa2258fd6afe57b